refactor: replace type interfaces with re-exports from model classes
Move PROFILE_IRIS to models/auth.ts and ProgrammingSubmissionState enum
to models/submissions.ts. Convert types/artemis.ts from 20 interface
definitions into a re-export barrel file, keeping only the
WebSocketMessageHandler callback interface locally.
