+++ iris-thaumantias/src/types/index.ts
@@ -8,4 +8,3 @@ export type {
     TrackedCourse,
     TrackedExercise
 } from './context';
-export { PROFILE_IRIS } from './artemis';
